# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ТЕСТОВОЕ API",
        "version": "1.0"
    },
    "paths": {
        "/api/order/create": {
            "get": {
                "tags": [
                    "Заказы"
                ],
                "description": "Создание заказа",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Почта"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Телефон"
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Адрес"
                    },
                    {
                        "name": "products",
                        "in": "query",
                        "description": "Продукты в формате JSON: {\"product_id\": count} где product_id - это id продукта, а count - это его количество. Пример: {\"1\": 3, \"2\": 5}"
                    },
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "API ключ"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/order/update": {
            "get": {
                "tags": [
                    "Заказы"
                ],
                "description": "Изменение заказа",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID Заказа"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Почта"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Телефон"
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Адрес"
                    },
                    {
                        "name": "products",
                        "in": "query",
                        "description": "Продукты в формате JSON: {\"product_id\": count} где product_id - это id продукта, а count - это его количество. Пример: {\"1\": 3, \"2\": 5}"
                    },
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "API ключ"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/order/read": {
            "get": {
                "tags": [
                    "Заказы"
                ],
                "description": "Информация о заказе",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID Заказа"
                    },
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "API ключ"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    },
                    "404": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/order/delete": {
            "get": {
                "tags": [
                    "Заказы"
                ],
                "description": "Удаление заказа",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID Заказа"
                    },
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "API ключ"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    },
                    "404": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Заказы"
                ],
                "description": "Список заказов",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы"
                    },
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Количество отображаемых заказов"
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Фильтрация заказов по дате (принимает любое значение)"
                    },
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "API ключ"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    },
                    "404": {
                        "description": "error"
                    }
                }
            }
        }
    }
}